Optimize releng cloning when running unit tests
not need to clone full releng
depth 1 is enough
do not go out of workspace anymore..never...

+++ run_unit_tests.sh
@@ -25,12 +25,13 @@ pip install -r $WORKSPACE/test-requirements.txt
 pip install $WORKSPACE
 
 #install releng
-cd $WORKSPACE/../
-rm -rf releng-unittests
-git clone https://gerrit.opnfv.org/gerrit/releng releng-unittests
+if [ -z releng-unittests ]
+then
+    rm -rf releng-unittests
+fi
+git clone --depth 1 https://gerrit.opnfv.org/gerrit/releng releng-unittests
 pip install releng-unittests/modules/
 rm -fr releng-unittests
-cd $WORKSPACE
 
 export CONFIG_FUNCTEST_YAML=$(pwd)/functest/ci/config_functest.yaml
 # unit tests
